Proper setValue implementation
The previous attempt at setting the output value was not complete.

Reworked the previous implementation and adapted everything the way it
should be. the new "setOutputValue" call was removed, the original setValue
call was fixed.

+++ core/model/addressablemodelitem.cpp
@@ -44,7 +44,7 @@ namespace dss {
     m_pApartment->getActionRequestInterface()->callScene(this, SceneDec);
   } // decreaseValue
 
-  void AddressableModelItem::setValue(const double _value) {
+  void AddressableModelItem::setValue(const uint8_t _value) {
     m_pApartment->getActionRequestInterface()->setValue(this, _value);
   } // setValue
